Add utility function to load configs
This is the start to a larger change. The first time you load up
alternaut, everything breaks because the autoload functions haven't been
defined yet. It's the core design challenge of using a function-based
API.

I'm migrating to a declarative config style instead, using
`g:alternaut#conventions` as the config key. This will replace
`alternaut#RegisterLanguage`.