Infer test role from file names
I've added some checks to the `is_test_file` function enabling alternaut
to more quickly (and accurately) determine if a file is a test file
purely by its name. Previously it was purely dependant on directory
structure.

This solves half of issue #5. Now the test and source files resolve, but
it relies on a highly expensive file query. I can do better.