Improve performance for co-located test searches
If you keep tests in the same directory as your source code, use the `.`
directory convention. It's treated specially by alternaut and only
searches the current directory for test/source files. Note that for that
directory convention, it bypasses upward search. It won't find the
source file if it's in a parent directory.